refactor: prepare bib dir for multiple files (#2)
This will scan that directory for all bib|org files and set it in the
appropriate variable for helm-bibtex to find. It DOES NOT update at
runtime, so if you add a file outside of helm-bibtex, then it won't pick
it up until you restart.

I have tried to create a file-watcher, but although the function
triggers, it does not update the list. This doesn't seem like the best
approach anyway.

Squashed commit of the following:

commit 023ab7fd994605cc78b98ef72673ef36e877451e
Author: xbc5 <87829033+xbc5@users.noreply.github.com>
Date:   Sat Jun 17 17:56:14 2023 +0100

    scan for bib files at boot

commit 522c55efc86ce078d36c72e07a04e4861e13099e
Author: xbc5 <87829033+xbc5@users.noreply.github.com>
Date:   Sat Jun 17 17:53:29 2023 +0100

    fix: update templates to reflect #38

    That commit changed the org irectory structure, but not the capture
    templates, so fix that.

commit 2e8f7cfc187563b71f7918b6f146a72b9a84dc95
Author: xbc5 <87829033+xbc5@users.noreply.github.com>
Date:   Sat Jun 17 17:22:31 2023 +0100

    update bib related global config var names